Initial Commit
+ Broke out CfgVehicles into individual component parts for readability and ease of use.
