+++ @AH-64D Apache Longbow/Addons/fza_ah64_controls/config/cfgVehicles.hpp
@@ -197,753 +197,15 @@ class CfgVehicles
 					position = "exhaust2";
 				};
 		};
-		class Sounds
-		{
-				class EngineExt
-				{
-					sound[] = {"\fza_ah64_US\audio\Engine", 1, 1.000000, 1200};
-					frequency = "rotorSpeed";
-					volume = "camPos*((rotorSpeed-0.72)*4)";
-				};
-				class RotorExt
-				{
-					sound[] = {"\fza_ah64_US\audio\Ext_Rotor.ogg", 1, 1.000000, 1200};
-					frequency = "rotorSpeed * (1 - rotorThrust/5)";
-					volume = "camPos*(0 max (rotorSpeed-0.1))*(1 + rotorThrust)";
-				};
-				class EngineClose
-				{
-					sound[] = {"\fza_ah64_US\audio\Engine_Close.ogg", 1, 1.000000, 1200};
-					frequency = "rotorSpeed";
-					volume = "camPos*((rotorSpeed-0.72)*4)";
-				};
-				class RotorClose
-				{
-					sound[] = {"\fza_ah64_US\audio\Rotor_Close.ogg", 1, 1.000000, 1200};
-					frequency = "rotorSpeed * (1 - rotorThrust/5)";
-					volume = "camPos*(0 max (rotorSpeed-0.1))*(1 + rotorThrust)";
-				};
-				class RotorSwist
-				{
-					sound[] = {"A3\Sounds_F\vehicles\air\Heli_Attack_01\swist", 0.500000, 1, 300};
-					frequency = 1;
-					volume = "camPos * (rotorThrust factor [0.7, 0.9])";
-				};
-				class TailRotor
-				{
-					cone[] = {3, 1.57, 3, 1.57};
-					frequency = "rotorSpeed";
-					sound[] = {"\fza_ah64_US\audio\Ext_Tail_Rotor.ogg", 3, 1, 1200};
-					volume = "camPos*(0 max (rotorSpeed-0.1))*(1 + rotorThrust)";
-				};
-				class Distance
-				{
-					sound[]  = {"\fza_ah64_US\audio\Engine_Far.ogg", 4, 1, 3500};
-					frequency = "rotorSpeed";
-					volume = "2 * camPos * (0 max (rotorSpeed-0.4))";
-				};
-				class FarDistance
-				{
-					sound[]  = {"\fza_ah64_US\audio\Rotor_Far.ogg", 4, 1, 3500};
-					frequency = "rotorSpeed";
-					volume = "2 * camPos * (0 max (rotorSpeed-0.4))";
-				};
-				class EngineInt
-				{
-					sound[] = {"\fza_ah64_us\audio\Int_Engine.ogg", 1, 1.000000};
-					frequency = "rotorSpeed";
-					volume = "(1-camPos)*((rotorSpeed-0.75)*4)";
-				};
-				class RotorInt
-				{
-					sound[] = {"\fza_ah64_us\audio\Int_Rotor.ogg", 1, 1};
-					frequency = "rotorSpeed * (1 - rotorThrust/5)";
-					volume = "(1-camPos)*(0 max (rotorSpeed-0.1))*(1 + rotorThrust)";
-				};
-				class TransmissionDamageExt_phase1
-				{
-					sound[] = {"A3\Sounds_F\vehicles\air\noises\heli_damage_transmission_ext_1", 1.000000, 1.000000, 150};
-					frequency = "0.66 + rotorSpeed / 3";
-					volume = "camPos * (transmissionDamage factor [0.3, 0.35]) * (transmissionDamage factor [0.5, 0.45]) * (rotorSpeed factor [0.2, 0.5])";
-				};
-				class TransmissionDamageExt_phase2
-				{
-					sound[] = {"A3\Sounds_F\vehicles\air\noises\heli_damage_transmission_ext_2", 1.000000, 1.000000, 150};
-					frequency = "0.66 + rotorSpeed / 3";
-					volume = "camPos * (transmissionDamage factor [0.45, 0.5]) * (rotorSpeed factor [0.2, 0.5])";
-				};
-				class TransmissionDamageInt_phase1
-				{
-					sound[] = {"A3\Sounds_F\vehicles\air\noises\heli_damage_transmission_int_1", 1.000000, 1.000000, 150};
-					frequency = "0.66 + rotorSpeed / 3";
-					volume = "(1 - camPos) * (transmissionDamage factor [0.3, 0.35]) * (transmissionDamage factor [0.5, 0.45]) * (rotorSpeed factor [0.2, 0.5])";
-				};
-				class TransmissionDamageInt_phase2
-				{
-					sound[] = {"A3\Sounds_F\vehicles\air\noises\heli_damage_transmission_int_2", 1.000000, 1.000000, 150};
-					frequency = "0.66 + rotorSpeed / 3";
-					volume = "(1 - camPos) * (transmissionDamage factor [0.45, 0.5]) * (rotorSpeed factor [0.2, 0.5])";
-				};
-				class rotorLowAlarmInt
-				{
-					sound[] = {"\fza_ah64_us\audio\betty\bt_rotorrpm_low.ogg", 1, 1.000000};
-					frequency = 1;
-					volume = "engineOn * (1 - camPos) * (rotorSpeed factor [0.9, 0.8999]) * (rotorSpeed factor [-0.5, 1]) * (speed factor [3, 3.01])";
-				};
-				class scrubLandInt
-				{
-					sound[] = {"A3\Sounds_F\vehicles\air\noises\wheelsInt", 1.000000, 1.000000, 100};
-					frequency = 1;
-					volume = "2 * (1-camPos) * (scrubLand factor[0.02, 0.05]) * (1 - (lateralMovement factor [0.7,1]))";
-				};
-				class scrubLandExt
-				{
-					sound[] = {"A3\Sounds_F\dummysound", 1.000000, 1.000000, 100};
-					frequency = 1;
-					volume = "camPos * (scrubLand factor[0.02, 0.05]) * (1 - (lateralMovement factor [0.7,1]))";
-				};
-				class scrubBuildingInt
-				{
-					sound[] = {"A3\Sounds_F\vehicles\air\noises\wheelsInt", 1.000000, 1.000000, 100};
-					frequency = 1;
-					volume = "(1-camPos) * (scrubBuilding factor[0.02, 0.05]) * (1 - (lateralMovement factor [0.7,1]))";
-				};
-				class scrubBuildingExt
-				{
-					sound[] = {"A3\Sounds_F\dummysound", 1.000000, 1.000000, 100};
-					frequency = 1;
-					volume = "camPos * (scrubBuilding factor[0.02, 0.05])";
-				};
-				class scrubTreeInt
-				{
-					sound[] = {"A3\Sounds_F\vehicles\air\noises\scrubTreeInt", 1.000000, 1.000000, 100};
-					frequency = 1;
-					volume = "(1 - camPos) * ((scrubTree) factor [0, 0.01])";
-				};
-				class scrubTreeExt
-				{
-					sound[] = {"A3\Sounds_F\vehicles\air\noises\scrubTreeExt", 1.000000, 1.000000, 100};
-					frequency = 1;
-					volume = "camPos * ((scrubTree) factor [0, 0.01])";
-				};
-				class RainExt
-				{
-					sound[] = {"\fza_ah64_us\audio\Ext_Rain.ogg", 1.000000, 1.000000, 100};
-					frequency = 1;
-					volume = "camPos * (rain - rotorSpeed/2) * 2";
-				};
-				class RainInt
-				{
-					sound[] = {"\fza_ah64_us\audio\Int_Rain.ogg", 1.000000, 1.000000, 100};
-					frequency = 1;
-					volume = "(1-camPos)*(rain - rotorSpeed/2)*2";
-				};
-				class SlingLoadDownExt
-				{
-					sound[] = {"A3\Sounds_F\vehicles\air\noises\SL_engineDownEXT", 1.258925, 1.000000, 500};
-					frequency = 1;
-					volume = "camPos*(slingLoadActive factor [0,-1])";
-				};
-				class SlingLoadUpExt
-				{
-					sound[] = {"A3\Sounds_F\vehicles\air\noises\SL_engineupEXT", 1.258925, 1.000000, 500};
-					frequency = 1;
-					volume = "camPos*(slingLoadActive factor [0,1])";
-				};
-				class SlingLoadDownInt
-				{
-					sound[] = {"A3\Sounds_F\vehicles\air\noises\SL_engineDownINT", 1.000000, 1.000000, 500};
-					frequency = 1;
-					volume = "(1-camPos)*(slingLoadActive factor [0,-1])";
-				};
-				class SlingLoadUpInt
-				{
-					sound[] = {"A3\Sounds_F\vehicles\air\noises\SL_engineUpINT", 1.000000, 1.000000, 500};
-					frequency = 1;
-					volume = "(1-camPos)*(slingLoadActive factor [0,1])";
-				};
-				class WindInt
-				{
-					sound[] = {"A3\Sounds_F\vehicles\air\noises\wind_closed", 0.398107, 1.000000, 50};
-					frequency = 1;
-					volume = "(1-camPos)*(speed factor[5, 50])*(speed factor[5, 50])";
-				};
-				class GStress
-				{
-					sound[] = {"A3\Sounds_F\vehicles\noises\vehicle_stress2b", 0.354813, 1.000000, 50};
-					frequency = 1;
-					volume = "engineOn * (1-camPos) * ((gmeterZ factor[1.5, 2.5]) + (gmeterZ factor[0.5, -0.5]))";
-				};
-		};
-		class SoundsExt
-		{
-			class Sounds
-			{
-				class EngineExt
-				{
-					sound[] = {"\fza_ah64_US\audio\Engine", 1, 1.000000, 1200};
-					frequency = "rotorSpeed";
-					volume = "camPos*((rotorSpeed-0.72)*4)";
-				};
-				class RotorExt
-				{
-					sound[] = {"\fza_ah64_US\audio\Ext_Rotor.ogg", 1, 1.000000, 1200};
-					frequency = "rotorSpeed * (1 - rotorThrust/5)";
-					volume = "camPos*(0 max (rotorSpeed-0.1))*(1 + rotorThrust)";
-				};
-				class EngineClose
-				{
-					sound[] = {"\fza_ah64_US\audio\Engine_Close.ogg", 1, 1.000000, 1200};
-					frequency = "rotorSpeed";
-					volume = "camPos*((rotorSpeed-0.72)*4)";
-				};
-				class RotorClose
-				{
-					sound[] = {"\fza_ah64_US\audio\Rotor_Close.ogg", 1, 1.000000, 1200};
-					frequency = "rotorSpeed * (1 - rotorThrust/5)";
-					volume = "camPos*(0 max (rotorSpeed-0.1))*(1 + rotorThrust)";
-				};
-				class RotorSwist
-				{
-					sound[] = {"A3\Sounds_F\vehicles\air\Heli_Attack_01\swist", 0.500000, 1, 300};
-					frequency = 1;
-					volume = "camPos * (rotorThrust factor [0.7, 0.9])";
-				};
-				class TailRotor
-				{
-					cone[] = {3, 1.57, 3, 1.57};
-					frequency = "rotorSpeed";
-					sound[] = {"\fza_ah64_US\audio\Ext_Tail_Rotor.ogg", 3, 1, 1200};
-					volume = "camPos*(0 max (rotorSpeed-0.1))*(1 + rotorThrust)";
-				};
-				class Distance
-				{
-					sound[]  = {"\fza_ah64_US\audio\Engine_Far.ogg", 4, 1, 3500};
-					frequency = "rotorSpeed";
-					volume = "2 * camPos * (0 max (rotorSpeed-0.4))";
-				};
-				class FarDistance
-				{
-					sound[]  = {"\fza_ah64_US\audio\Rotor_Far.ogg", 4, 1, 3500};
-					frequency = "rotorSpeed";
-					volume = "2 * camPos * (0 max (rotorSpeed-0.4))";
-				};
-				class EngineInt
-				{
-					sound[] = {"\fza_ah64_us\audio\Int_Engine.ogg", 1, 1.000000};
-					frequency = "rotorSpeed";
-					volume = "(1-camPos)*((rotorSpeed-0.75)*4)";
-				};
-				class RotorInt
-				{
-					sound[] = {"\fza_ah64_us\audio\Int_Rotor.ogg", 1, 1};
-					frequency = "rotorSpeed * (1 - rotorThrust/5)";
-					volume = "(1-camPos)*(0 max (rotorSpeed-0.1))*(1 + rotorThrust)";
-				};
-				class TransmissionDamageExt_phase1
-				{
-					sound[] = {"A3\Sounds_F\vehicles\air\noises\heli_damage_transmission_ext_1", 1.000000, 1.000000, 200};
-					frequency = "0.66 + rotorSpeed / 3";
-					volume = "camPos * (transmissionDamage factor [0.3, 0.35]) * (transmissionDamage factor [0.5, 0.45]) * (rotorSpeed factor [0.2, 0.5])";
-				};
-				class TransmissionDamageExt_phase2
-				{
-					sound[] = {"A3\Sounds_F\vehicles\air\noises\heli_damage_transmission_ext_2", 1.000000, 1.000000, 200};
-					frequency = "0.66 + rotorSpeed / 3";
-					volume = "camPos * (transmissionDamage factor [0.45, 0.5]) * (rotorSpeed factor [0.2, 0.5])";
-				};
-				class TransmissionDamageInt_phase1
-				{
-					sound[] = {"A3\Sounds_F\vehicles\air\noises\heli_damage_transmission_int_1", 1.000000, 1.000000, 200};
-					frequency = "0.66 + rotorSpeed / 3";
-					volume = "(1 - camPos) * (transmissionDamage factor [0.3, 0.35]) * (transmissionDamage factor [0.5, 0.45]) * (rotorSpeed factor [0.2, 0.5])";
-				};
-				class TransmissionDamageInt_phase2
-				{
-					sound[] = {"A3\Sounds_F\vehicles\air\noises\heli_damage_transmission_int_2", 1.000000, 1.000000, 200};
-					frequency = "0.66 + rotorSpeed / 3";
-					volume = "(1 - camPos) * (transmissionDamage factor [0.45, 0.5]) * (rotorSpeed factor [0.2, 0.5])";
-				};
-			};
-		};
 
-		#define HIT_RADIUS_01 0.14
-		#define HIT_RADIUS_02 0.05
-		#define ARMOR_SCALAR  0.067
+		#include "cfgVehicles\sounds.hpp"
+		#include "cfgVehicles\soundsExt.hpp"
+		#include "cfgVehicles\userActions.hpp"
 
-		class HitPoints		
-		{		
-			class hit_apu	
-			{	
-				armor = 1.44 * ARMOR_SCALAR;
-				radius = HIT_RADIUS_01;
-				minimalHit = 0.1;
-				explosionShielding = 0.30;
-				name = "hit_apu";
-				material = 51;
-				passThrough = 0;
-			};	
-			class hit_drives_intermediategearbox	
-			{	
-				armor = 1.62 * ARMOR_SCALAR;
-				radius = HIT_RADIUS_01;
-				minimalHit = 0.1;
-				explosionShielding = 0.34;
-				name = "hit_drives_intermediategearbox";
-				material = 51;
-				passThrough = 0;
-			};	
-			class hit_drives_noseGearbox1	
-			{	
-				armor = 1.44 * ARMOR_SCALAR;
-				radius = HIT_RADIUS_01;
-				minimalHit = 0.1;
-				explosionShielding = 0.24;
-				name = "hit_drives_noseGearbox1";
-				material = 51;
-				passThrough = 0;
-			};	
-			class hit_drives_noseGearbox2	
-			{	
-				armor = 1.44 * ARMOR_SCALAR;
-				radius = HIT_RADIUS_01;
-				minimalHit = 0.1;
-				explosionShielding = 0.24;
-				name = "hit_drives_noseGearbox2";
-				material = 51;
-				passThrough = 0;
-			};	
-			class hit_drives_tailrotorgearbox	
-			{	
-				armor = 1.62 * ARMOR_SCALAR;
-				radius = HIT_RADIUS_01;
-				minimalHit = 0.1;
-				explosionShielding = 0.34;
-				name = "hit_drives_tailrotorgearbox";
-				material = 51;
-				passThrough = 0;
-			};	
-			class hit_drives_transmission	
-			{	
-				armor = 2.88 * ARMOR_SCALAR;
-				radius = HIT_RADIUS_01;
-				minimalHit = 0.1;
-				explosionShielding = 0.30;
-				name = "hit_drives_transmission";
-				material = 51;
-				passThrough = 0;
-			};	
-			class hit_elec_battery	
-			{	
-				armor = 1.44 * ARMOR_SCALAR;
-				radius = HIT_RADIUS_02;
-				minimalHit = 0.1;
-				explosionShielding = 0.24;
-				name = "hit_elec_battery";
-				material = 51;
-				passThrough = 0;
-			};	
-			class hit_elec_generator1	
-			{	
-				armor = 0.72 * ARMOR_SCALAR;
-				radius = HIT_RADIUS_02;
-				minimalHit = 0.1;
-				explosionShielding = 0.09;
-				name = "hit_elec_generator1";
-				material = 51;
-				passThrough = 0;
-			};	
-			class hit_elec_generator2	
-			{	
-				armor = 0.72 * ARMOR_SCALAR;
-				radius = HIT_RADIUS_02;
-				minimalHit = 0.1;
-				explosionShielding = 0.09;
-				name = "hit_elec_generator2";
-				material = 51;
-				passThrough = 0;
-			};
-			class hit_elec_rectifier1
-			{
-				armor = 0.72 * ARMOR_SCALAR;
-				radius = HIT_RADIUS_02;
-				minimalHit = 0.1;
-				explosionShielding = 0.10;
-				name = "hit_elec_rectifier1";
-				material = 51;
-				passThrough = 0;
-			};
-			class hit_elec_rectifier2
-			{
-				armor = 0.72 * ARMOR_SCALAR;
-				radius = HIT_RADIUS_02;
-				minimalHit = 0.1;
-				explosionShielding = 0.10;
-				name = "hit_elec_rectifier2";
-				material = 51;
-				passThrough = 0;
-			};
-			class hitengine1	
-			{	
-				armor = 1.80 * ARMOR_SCALAR;
-				radius = HIT_RADIUS_01;
-				minimalHit = 0.1;
-				explosionShielding = 0.30;
-				name = "hitengine1";
-				material = 51;
-				passThrough = 0;
-			};	
-			class hitengine2	
-			{	
-				armor = 1.80 * ARMOR_SCALAR;
-				radius = HIT_RADIUS_01;
-				minimalHit = 0.1;
-				explosionShielding = 0.30;
-				name = "hitengine2";
-				material = 51;
-				passThrough = 0;
-			};
-			class hit_fuel_aft {
-				armor = 1.44 * ARMOR_SCALAR;
-				radius = HIT_RADIUS_01;
-				minimalHit = 0.1;
-				explosionShielding = 0.15;
-				name = "hit_fuel_aft";
-				material = 51;
-				passThrough = 0;
-			};
-			class hit_fuel_forward {
-				armor = 1.44 * ARMOR_SCALAR;
-				radius = HIT_RADIUS_01;
-				minimalHit = 0.1;
-				explosionShielding = 0.15;
-				name = "hit_fuel_forward";
-				material = 51;
-				passThrough = 0;
-			};
-			class hit_msnEquip_magandrobbie {
-				armor = 1.44 * ARMOR_SCALAR;
-				radius = HIT_RADIUS_01;
-				minimalHit = 0.1;
-				explosionShielding = 0.15;
-				name = "hit_msnEquip_magandrobbie";
-				material = 51;
-				passThrough = 0;
-			};
-			class hit_hyd_prireservoir	
-			{	
-				armor = 0.72 * ARMOR_SCALAR;
-				radius = HIT_RADIUS_02;
-				minimalHit = 0.1;
-				explosionShielding = 0.20;
-				name = "hit_hyd_prireservoir";
-				material = 51;
-				passThrough = 0;
-			};	
-			class hit_hyd_priPump	
-			{	
-				armor = 0.90 * ARMOR_SCALAR;
-				radius = HIT_RADIUS_02;
-				minimalHit = 0.1;
-				explosionShielding = 0.15;
-				name = "hit_hyd_priPump";
-				material = 51;
-				passThrough = 0;
-			};	
-			class hit_hyd_utilreservoir	
-			{	
-				armor = 0.72 * ARMOR_SCALAR;
-				radius = HIT_RADIUS_02;
-				minimalHit = 0.1;
-				explosionShielding = 0.20;
-				name = "hit_hyd_utilreservoir";
-				material = 51;
-				passThrough = 0;
-			};	
-			class hit_hyd_utilPump	
-			{	
-				armor = 0.90 * ARMOR_SCALAR;
-				radius = HIT_RADIUS_02;
-				minimalHit = 0.1;
-				explosionShielding = 0.15;
-				name = "hit_hyd_utilPump";
-				material = 51;
-				passThrough = 0;
-			};
-			class hithrotor
-			{
-				armor = 5.76 * ARMOR_SCALAR;
-				radius = HIT_RADIUS_01;
-				minimalHit = 0.1;
-				explosionShielding = 0.80;
-				name = "hithrotor";
-				material = 51;
-				passThrough = 0;
-			};
-			class hit_msnEquip_gun_turret	
-			{	
-				armor = 1.44 * ARMOR_SCALAR;
-				radius = HIT_RADIUS_01;
-				minimalHit = 0.1;
-				explosionShielding = 0.80;
-				name = "hit_msnEquip_gun_turret";
-				material = 51;
-				passThrough = 0;
-			};	
-			class hit_msnEquip_pnvs_flir	
-			{	
-				armor = 1.44 * ARMOR_SCALAR;
-				radius = HIT_RADIUS_01;
-				minimalHit = 0.1;
-				explosionShielding = 0.80;
-				name = "hit_msnEquip_pnvs_flir";
-				material = 51;
-				passThrough = 0;
-			};	
-			class hit_msnEquip_pnvs_turret	
-			{	
-				armor = 1.44 * ARMOR_SCALAR;
-				radius = HIT_RADIUS_01;
-				minimalHit = 0.1;
-				explosionShielding = 0.80;
-				name = "hit_msnEquip_pnvs_turret";
-				material = 51;
-				passThrough = 0;
-			};	
-			class hit_msnEquip_pylon1	
-			{	
-				armor = 1.44 * ARMOR_SCALAR;
-				radius = HIT_RADIUS_01;
-				minimalHit = 0.1;
-				explosionShielding = 0.80;
-				name = "hit_msnEquip_pylon1";
-				material = 51;
-				passThrough = 0;
-			};	
-			class hit_msnEquip_pylon2	
-			{	
-				armor = 1.44 * ARMOR_SCALAR;
-				radius = HIT_RADIUS_01;
-				minimalHit = 0.1;
-				explosionShielding = 0.80;
-				name = "hit_msnEquip_pylon2";
-				material = 51;
-				passThrough = 0;
-			};	
-			class hit_msnEquip_pylon3	
-			{	
-				armor = 1.44 * ARMOR_SCALAR;
-				radius = HIT_RADIUS_01;
-				minimalHit = 0.1;
-				explosionShielding = 0.80;
-				name = "hit_msnEquip_pylon3";
-				material = 51;
-				passThrough = 0;
-			};	
-			class hit_msnEquip_pylon4	
-			{	
-				armor = 1.44 * ARMOR_SCALAR;
-				radius = HIT_RADIUS_01;
-				minimalHit = 0.1;
-				explosionShielding = 0.80;
-				name = "hit_msnEquip_pylon4";
-				material = 51;
-				passThrough = 0;
-			};	
-			class hit_msnEquip_tads_dtv	
-			{	
-				armor = 1.44 * ARMOR_SCALAR;
-				radius = HIT_RADIUS_01;
-				minimalHit = 0.1;
-				explosionShielding = 0.80;
-				name = "hit_msnEquip_tads_dtv";
-				material = 51;
-				passThrough = 0;
-			};	
-			class hit_msnEquip_tads_flir	
-			{	
-				armor = 1.44 * ARMOR_SCALAR;
-				radius = HIT_RADIUS_01;
-				minimalHit = 0.1;
-				explosionShielding = 0.80;
-				name = "hit_msnEquip_tads_flir";
-				material = 51;
-				passThrough = 0;
-			};	
-			class hit_msnEquip_tads_turret	
-			{	
-				armor = 1.44 * ARMOR_SCALAR;
-				radius = HIT_RADIUS_01;
-				minimalHit = 0.1;
-				explosionShielding = 0.80;
-				name = "hit_msnEquip_tads_turret";
-				material = 51;
-				passThrough = 0;
-			};	
-			class hit_msnEquip_fcr	
-			{	
-				armor = 1.44 * ARMOR_SCALAR;
-				radius = HIT_RADIUS_01;
-				minimalHit = 0.1;
-				explosionShielding = 0.80;
-				name = "hit_msnEquip_fcr";
-				material = 51;
-				passThrough = 0;
-			};	
-			class hit_stabilator	
-			{	
-				armor = 1.44 * ARMOR_SCALAR;
-				radius = HIT_RADIUS_01;
-				minimalHit = 0.1;
-				explosionShielding = 0.80;
-				name = "hit_stabilator";
-				material = 51;
-				passThrough = 0;
-			};
-			class hitvrotor
-			{	
-				armor = 5.76 * ARMOR_SCALAR;
-				radius = HIT_RADIUS_01;
-				minimalHit = 0.1;
-				explosionShielding = 0.80;
-				name = "hitvrotor";
-				material = 51;
-				passThrough = 0;
-			};		
-		};		
-		class UserActions
-		{
-			class lookcenter
-			{
-				displayName="";
-				useAction=false;
-				showSwitchAction=false;
-				position="pilot_action";
-				onlyForPlayer=1;
-				radius=8;
-				showWindow=0;
-				priority=0;
-				condition="(player == driver this || player == gunner this) && alive this && !(this iskindof ""fza_ah64a_l"")";
-				shortcut="LookCenter";
-				statement="fza_ah64_mousehorpos = 0.5; fza_ah64_mousevertpos = 0.5; ((uiNameSpace getVariable 'fza_ah64_click_helper') displayCtrl 601) ctrlSetPosition[fza_ah64_mousehorpos - 0.005, fza_ah64_mousevertpos - 0.009]; ((uiNameSpace getVariable 'fza_ah64_click_helper') displayCtrl 602) ctrlSetPosition[fza_ah64_mousehorpos - 0.25, fza_ah64_mousevertpos + 0.02]; ((uiNameSpace getVariable 'fza_ah64_click_helper') displayCtrl 601) ctrlCommit 0.01; ((uiNameSpace getVariable 'fza_ah64_click_helper') displayCtrl 602) ctrlCommit 0.01;";
-			};
-			class sensorselect
-			{
-				displayName="";
-				useAction=false;
-				showSwitchAction=false;
-				position="pilot_action";
-				onlyForPlayer=1;
-				radius=8;
-				showWindow=0;
-				priority=0;
-				//condition="(player == driver this || player == gunner this) && alive this && !(this iskindof ""fza_ah64a_l"" || this iskindof ""fza_ah64a_e"")";
-				condition="(player == driver this || player == gunner this) && alive this && !(this iskindof ""fza_ah64a_l"")";
-				shortcut="OpticsMode";
-				statement="[this] call fza_fnc_weaponguncontrol;";
-			};
-			class tiron
-			{
-				displayName="Toggle Head Tracking On";
-				useAction=false;
-				showSwitchAction=false;
-				position="pilot_action";
-				onlyForPlayer=1;
-				radius=8;
-				showWindow=0;
-				priority=9;
-				condition="!fza_ah64_tiron && (player == driver this || player == gunner this)";
-				statement="fza_ah64_tiron = !fza_ah64_tiron";
-			};
-			class tiroff
-			{
-				displayName="Toggle Head Tracking Off";
-				useAction=false;
-				showSwitchAction=false;
-				position="pilot_action";
-				onlyForPlayer=1;
-				radius=8;
-				showWindow=0;
-				priority=9;
-				condition="fza_ah64_tiron && (player == driver this || player == gunner this)";
-				statement="fza_ah64_tiron = !fza_ah64_tiron";
-			};
-		};
 		class MFD {};
 		
-		class AcreIntercoms
-		{
-            class Intercom_1
-			{
-                displayName = "Intercom (ICS)";
-                shortName = "ICS";
-                allowedPositions[] = {"crew"};
-                disabledPositions[] = {};
-                masterPositions[] = {"crew"}; // Seats with master stations have the possibility of broadcasting a message in that network
-                limitedPositions[] = {"cargo"}; // Despite not having regular access to the network, units in cargo positions can have limited connections to communicate with the crew. These positions do not transmit automatically in the limited network; units in this position must toggle the functionality manually.
-                numLimitedPositions = 3;  // This is the number of simultaneous connections that units defined in the previous array can have
-                connectedByDefault = 0; // The intercom initial configuration is enabled upon entering a vehicle
-            };
-		};
-		
-		acre_hasInfantryPhone = 1; // 1 - enabled, 0 - disabled
-		acre_infantryPhoneDisableRinging = 1;   // If set to 1, the ringing funtionality will not be available
-		acre_infantryPhoneCustomRinging[] = {}; // An array used in order to override the default sound for the ringing functionality
-		acre_infantryPhoneIntercom[] = {"intercom_1"}; // List of intercom names (intercom_1, intercom_2) or "all" in order to specify which intercom networks the phone can connect to
-		acre_infantryPhoneControlActions[] = {"intercom_1"}; // Only those units in "intercom_1" can have access to ringing functionality
-		acre_eventInfantryPhone = QFUNC(noApiFunction);
-		
-       class AcreRacks
-        {
-            class Rack_1
-            {
-               displayName = "ARC-186 (VHF)"; // Name is displayed in the interaction menu.
-               shortName = "VHF";
-               allowedPositions[] = {"driver","gunner"};     // Who can configure the radio and open the radio GUI. Same wildcards as the intercom. It also allows transmitting/receiving
-               componentName = "ACRE_VRC103";
-               disabledPositions[] = {};
-               defaultComponents[] = {};                     // Use this to attach simple components like Antennas. Not yet fully implemented
-               mountedRadio = "ACRE_PRC117F";                 // Predefined mounted radio
-               isRadioRemovable = 0;
-               intercom[] = {};                   // All units in intercom will be able to hear/send transmittions (ACE3 interaction menu) but they cannot manipulate the radio (GUI interface)
-           };
-            class Rack_2
-            {
-               displayName = "ARC-164 (UHF)"; // Name is displayed in the interaction menu.
-               shortName = "UHF";
-               allowedPositions[] = {"driver","gunner"};     // Who can configure the radio and open the radio GUI. Same wildcards as the intercom. It also allows transmitting/receiving
-               componentName = "ACRE_VRC103";
-               disabledPositions[] = {};
-               defaultComponents[] = {};                     // Use this to attach simple components like Antennas. Not yet fully implemented
-               mountedRadio = "ACRE_PRC117F";                 // Predefined mounted radio
-               isRadioRemovable = 0;
-               intercom[] = {};                   // All units in intercom will be able to hear/send transmittions (ACE3 interaction menu) but they cannot manipulate the radio (GUI interface)
-           };
-			class Rack_3
-            {
-               displayName = "ARC-201D (FM1)"; // Name is displayed in the interaction menu.
-               shortName = "FM1";
-               allowedPositions[] = {"driver","gunner"};     // Who can configure the radio and open the radio GUI. Same wildcards as the intercom. It also allows transmitting/receiving
-               componentName = "ACRE_SEM90";
-               disabledPositions[] = {};
-               defaultComponents[] = {};                     // Use this to attach simple components like Antennas. Not yet fully implemented
-               mountedRadio = "ACRE_SEM70";                 // Predefined mounted radio
-               isRadioRemovable = 0;
-               intercom[] = {};                   // All units in intercom will be able to hear/send transmittions (ACE3 interaction menu) but they cannot manipulate the radio (GUI interface)
-           };
-            class Rack_4
-            {
-               displayName = "ARC-201D (FM2)"; // Name is displayed in the interaction menu.
-               shortName = "FM2";
-               allowedPositions[] = {"driver","gunner"};     // Who can configure the radio and open the radio GUI. Same wildcards as the intercom. It also allows transmitting/receiving
-               componentName = "ACRE_SEM90";
-               disabledPositions[] = {};
-               defaultComponents[] = {};                     // Use this to attach simple components like Antennas. Not yet fully implemented
-               mountedRadio = "ACRE_SEM70";                 // Predefined mounted radio
-               isRadioRemovable = 0;
-               intercom[] = {};                   // All units in intercom will be able to hear/send transmittions (ACE3 interaction menu) but they cannot manipulate the radio (GUI interface)
-           };
-        };
+		#include "cfgVehicles\acre.hpp"
+
 		class Attributes
 		{
 			class fza_pylonLoadout
@@ -1075,298 +337,15 @@ class CfgVehicles
 		envelope[]={
 			0.0,0.6,1.6,3.2,3.8,5.0,5.25,5.4,5.6,5.7,5.8,5.9,6.0,4.0,1.5 // lift
 		};
-		class textureSources
-		{
-			class b2
-			{
-				displayName = "Default";
-				author = "NodUnit";
-				textures[]= {"\fza_ah64_us\tex\Ex\b2_co.paa","\fza_ah64_us\tex\ex\fcr_co.paa"};
-			};
-			class b2_weather
-			{
-				displayName = "Default (weathered)";
-				author = "Franze";
-				textures[]= {"\fza_ah64_us\tex\Ex\b2_weather_co.paa","\fza_ah64_us\tex\ex\fcr_co.paa"};
-			};
-			class arb229th
-			{
-				displayName = "229th ARB";
-				author = "Apache mod development team";
-				textures[]= {"\fza_ah64_us\tex\Ex\229arb_co.paa","\fza_ah64_us\tex\ex\fcr_co.paa"};
-			};
-			class arb229th_weather
-			{
-				displayName = "229th ARB (weathered)";
-				author = "Apache mod development team";
-				textures[]= {"\fza_ah64_us\tex\Ex\229arb_weather_co.paa","\fza_ah64_us\tex\ex\fcr_co.paa"};
-			};
-			class indian_air_force
-			{
-				displayName = "Indian Air Force";
-				author = "BradMick";
-				textures[]= {"\fza_ah64_us\tex\Ex\b2_india_co.paa","\fza_ah64_us\tex\ex\fcr_co.paa"};
-			};
-			class rnlaf_q_09
-			{
-				displayName = "RNLAF 301 Squadron Q-09";
-				author = "Houseman";
-				textures[]= {"\fza_ah64_us\tex\Ex\b2_rnlaf_q_09_co.paa","\fza_ah64_us\tex\ex\fcr_co.paa"};
-			};
-			class rnlaf_demo
-			{
-				displayName = "RNLAF 301 Squadron Demo Team";
-				author = "Houseman";
-				textures[]= {"\fza_ah64_us\tex\Ex\b2_rnlaf_demo_co.paa","\fza_ah64_us\tex\ex\fcr_co.paa"};
-			};
-			class UK_AAC
-			{
-				displayName = "United Kingdom Army Air Corps";
-				author = "Rosd6(Dryden) & NodUnit";
-				textures[]= {"\fza_ah64_us\tex\Ex\UKAAC.paa","\fza_ah64_us\tex\ex\fcr_co.paa"};
-			};
-			class UK_AAC_weathered
-			{
-				displayName = "United Kingdom Army Air Corps (weathered)";
-				author = "Rosd6(Dryden) & NodUnit";
-				textures[]= {"\fza_ah64_us\tex\Ex\UKAAC_weathered.paa","\fza_ah64_us\tex\ex\fcr_co.paa"};
-			};
-			class UK_AAC_SF
-			{
-				displayName = "United Kingdom Army Air Corps Special Forces";
-				author = "Rosd6(Dryden) & NodUnit";
-				textures[]= {"\fza_ah64_us\tex\Ex\UKAAC_SF.paa","\fza_ah64_us\tex\ex\fcr_co.paa"};
-			};
-			class 211th_clean
-			{
-				displayName = "1/211th ARB ""The Air Pirates"" Utah National Guard";
-				author = "seven10 & Apache mod development team";
-				textures[]= {"\fza_ah64_us\tex\Ex\211th_co.paa","\fza_ah64_us\tex\ex\fcr_co.paa"};
-			};
-			class 211th_weathered
-			{
-				displayName = "1/211th ARB ""The Air Pirates"" Utah National Guard (weathered)";
-				author = "seven10 & Apache mod development team";
-				textures[]= {"\fza_ah64_us\tex\Ex\211th_weather_co.paa","\fza_ah64_us\tex\ex\fcr_co.paa"};
-			};
-			class Israeli_Air_Force
-			{
-				displayName = "Israeli Air Force (IAF Saraf)";
-				author = "NodUnit";
-				textures[]= {"\fza_ah64_us\tex\Ex\Nodunit_Isreal_co.paa","\fza_ah64_us\tex\ex\fcr_co.paa"};
-			};
-			class Shield
-			{
-				displayName = "Shield (Marvel)";
-				author = "NodUnit";
-				textures[]= {"\fza_ah64_us\tex\Ex\Shield.paa","\fza_ah64_us\tex\ex\fcr_co.paa"};
-			};
-			class NodUnit
-			{
-				displayName = "Kane's Wrath";
-				author = "NodUnit";
-				textures[]= {"\fza_ah64_us\tex\Ex\Nodunit.paa","\fza_ah64_us\tex\ex\fcr_co.paa"};
-			};
-		};
-		textureList[] = {"b2", 1};
-		hiddenSelectionsTextures[] = {"\fza_ah64_us\tex\Ex\b2_co.paa","\fza_ah64_us\tex\ex\fcr_co.paa"};
-		class Turrets
-		{
-			class MainTurret: NewTurret
-			{
-				gunnerAction = "fza_ah64_copilot";
-				gunnerGetInAction = "pilot_Heli_Light_02_Enter";
-				gunnerGetOutAction = "GetOutHigh";
-				preciseGetInOut = 1;
-				hideWeaponsGunner = false;
-				primary = 1;
-				primaryGunner = 1;
-				stabilizedInAxes = 3;
-				weapons[] = {"fza_ma_safe", "fza_gun_safe", "fza_rkt_safe", "fza_msl_safe", "Laserdesignator_mounted", "fza_burstlimiter","fza_m230"};
-				magazines[] = {"fza_safe", "LaserBatteries", "fza_m230_300"};
-				memoryPointsGetInGunner = "pos gunner";
-			    memoryPointsGetInGunnerDir = "pos gunner dir";
-			    memoryPointGun = "laserBegin";
-				memoryPointGunnerOptics = "gunnerview";
-				body = "tads_tur";
-				gun = "tads";
-				animationsourcebody = "tads_tur";
-				animationsourcegun = "tads";
-				gunBeg = "laserBegin";
-				gunEnd = "laserEnd";
-				gunnerOpticsModel = "";
-				gunnerOpticsColor[] = {1,1,1,1};
-				minElev = -60;
-				maxElev = 30;
-				initElev= 0;
-				maxXRotSpeed = 1;	// yawing speed
-				maxYRotSpeed = 1;	// pitching speed
-				maxMouseXRotSpeed= 0.5;
-				maxMouseYRotSpeed= 0.5;
-				minTurn = -120;
-				maxTurn = 120;
-				initTurn = 0;
-				minGunElev = -60;
-				maxGunElev = 30;
-				minGunTurn = -120;
-				maxGunTurn = 120;
-				minGunTurnAI = -90;
-				maxGunTurnAI = 90;
-				commanding = -1;
-				gunnerForceOptics = 0;
-				startEngine=0;
-				outGunnerMayFire = 1;
-				turretinfotype = "RscUnitInfoNoHUD";
-				turretFollowFreeLook = 0;
-				discreteDistance[]={100,200,300,400,500,600,700,800,1000,1200,1500,1800,2100,2500,2800,3300};
-				discreteDistanceInitIndex=5;
-				isCopilot = 1;
-				usePiP=1;
-				class Reflectors
-				{
-					class cabin
-					{
-						color[]={0.306, 0.878, 0.349};
-						ambient[] = {0.306, 0.878, 0.349};
-						intensity = 25;
-						size = 1;
-						innerAngle = 30;
-						outerAngle = 150;
-						coneFadeCoef = 1;
-						position = "plt_floodlamps";
-						direction = "plt_memflood";
-						hitpoint = "plt_floodlamps";
-						selection = "plt_floodlamps";
-						useFlare = 0;
-						flareSize = 0;
-						flareMaxDistance = 0;
-						dayLight = 1;
-						blinking = 0;
-						class Attenuation
-						{
-							start           = 0;
-							constant        = 0;
-							linear          = 1;
-							quadratic       = 2;
-							hardLimitStart  = 0.65;
-							hardLimitEnd    = 1.9;
-						};
-					};
-					class cargo_light_1: cabin
-					{
-						position = "cpg_flood";
-						direction = "cpg_memflood";
-						hitpoint = "cpg_flood";
-						selection = "cpg_flood";
-					};
-				};
-				class HitPoints
-				{
-					class HitTurret
-					{
-						armor = 0.9;
-						material = 51;
-						name = "tads_tur";
-						visual = "skin_tads1";
-						passThrough = 1;
-					};
-					class HitGun
-					{
-						armor = 1.3;
-						material = 52;
-						name = "otochlaven";
-						visual = "skin_otochlaven";
-						passThrough = 1;
-					};
-				};
-				class OpticsIn
-				{
-					class Wide
-					{
-						gunneropticsmodel = "\fza_ah64_us\fza_ah64_optics_empty";
-						directionStabilized = 1;
-						initanglex = 0;
-						initangley = 0;
-						initfov = 0.466;
-						maxanglex = 30;
-						maxangley = 120;
-						maxfov = 0.466;
-						minanglex = -60;
-						minangley = -120;
-						minfov = 0.466;
-						opticsdisplayname = "W";
-						thermalmode[] = {0,1};
-						visionmode[] = {"Normal","Ti"};
-					};
-					class Medium: Wide
-					{
-						gunneropticsmodel = "\fza_ah64_us\fza_ah64_optics_empty";
-						initfov = 0.09;
-						maxfov = 0.09;
-						minfov = 0.09;
-						opticsdisplayname = "M";
-					};
-					class Narrow: Wide
-					{
-						gunneropticsmodel = "\fza_ah64_us\fza_ah64_optics_empty";
-						initfov = 0.01;
-						maxfov = 0.01;
-						minfov = 0.01;
-						opticsdisplayname = "N";
-					};
-				};
-				class OpticsOut
-				{
-					class Monocular //FOV CPG
-					{
-						gunneropticseffect[] = {};
-						gunneropticsmodel = "";
-						initfov = 0.7;
-						initanglex = 0;
-						initangley = 0;
-						minfov = 0.4;
-						maxfov = 0.7;
-						minanglex = -60;
-						maxanglex = 30;
-						minangley = -120;
-						maxangley = 120;
-					};
-				};
-				class Components
-				{
-					class VehicleSystemsDisplayManagerComponentLeft
-					{
-						componentType = "VehicleSystemsDisplayManager";
-						x = "(profilenamespace getvariable [""IGUI_GRID_CUSTOMINFOLEFT_X"",	(safezoneX + 0.5 * (((safezoneW / safezoneH) min 1.2) / 40))])";
-						y = "(profilenamespace getvariable [""IGUI_GRID_CUSTOMINFOLEFT_Y"",	(safezoneY + safezoneH - 21 * ((((safezoneW / safezoneH) min 1.2) / 1.2) / 25))])";
-						left = 1;
-						defaultDisplay = "CrewDisplay";
-						class Components
-						{
-							class EmptyDisplay		// Empty display - hide panel
-							{
-								componentType = "EmptyDisplayComponent";
-							};
-							class MinimapDisplay	// GPS
-							{
-								componentType = "MinimapDisplayComponent";
-								resource = "RscCustomInfoMiniMap";
-							};
-						};
-					};
-					class VehicleSystemsDisplayManagerComponentRight : VehicleSystemsDisplayManagerComponentLeft {
-						left = 0;
-						right = 1;
-						x = "(profilenamespace getvariable [""IGUI_GRID_CUSTOMINFORIGHT_X"",((safezoneX + safezoneW) - ((10 * (((safezoneW / safezoneH) min 1.2) / 40)) + 0.5 * (((safezoneW / safezoneH) min 1.2) / 40)))])";
-						y = "(profilenamespace getvariable [""IGUI_GRID_CUSTOMINFORIGHT_Y"",(safezoneY + safezoneH - 21 * ((((safezoneW / safezoneH) min 1.2) / 1.2) / 25))])";
-					};
-				};
-			};
-		};
+
+		#include "cfgVehicles\textureSources.hpp"
+		#include "cfgVehicles\turrets.hpp"
+
 		class Library
 		{
 			libTextDesc = "The AH-64D Block II is an anti-tank attack helicopter. Defining characteristics are: Four bladed main rotor, X-shaped tail rotor, a boxy cockpit with two distinct bays on the sides, two stub wings with two pylons each, two main wheels and a tail wheel, a chin mounted turret, and a long rounded tailboom with a stabilizator at the rear. Production began in 1982 and the first unit was deployed in 1986. By 1993, the US Army had over 800 Apaches in service. It was firsted used in combat during Operation Just Cause, the invasion of Panama in 1989 where some drawbacks were encountered (primarily with the avionics). During Desert Storm in 1991, Apaches performed the first strike of that conflict by destroying Iraqi early warning radars. More drawbacks were encountered in Desert Storm, some of which were: limited navigation capabilities, high maintenance, and intolerance to desert conditions. AH-64As went through some modernization upgrades, but the AH-64A has been phased out of active US Army service in favor of the AH-64D, and now only National Guard and Reserve units operate the A model. The AH-64D has two General Electric T700-GE-701C engines producing a total of 3392shp, giving the aircraft a maximum clean speed of about 227mph (365kmh). This speed comes down to about 300kmh loaded. Combat endurance is about 2 hours with standard loads. It is armed with a M230 30mm Chain Gun cannon by default, with provisions for AGM-114 HELLFIRE missiles and 2.75in HYDRA-70 rockets. Experimentally it has used the Stinger missile, Sidearm missile, Sidewinder missile, and Starstreak missile. The US Army has not used any of these weapons operationally with the Apache.";
 		};
+
 		class HeadLimits
 		{
 			initanglex = 0;
@@ -1416,565 +395,9 @@ class CfgVehicles
 			visionmode[] = {"Ti"}; //PNVS
 		};
 
-		class UVAnimations
-		{
-			class plt_mmap_01_tnsx
-			{
-				type			= translation;
-				source			= plt_mmap_tx;
-				section			= plt_pr_mpd_back;
-				minValue		= -1;
-				maxValue		= 2;
-				offset0[]		= {-1,0};
-				offset1[]		= {2,0};
-			};
-			class plt_mmap_02_tnsy
-			{
-				type			= translation;
-				source			= plt_mmap_ty;
-				section			= plt_pr_mpd_back;
-				minValue		= -1;
-				maxValue		= 2;
-				offset0[]		= {0,-1};
-				offset1[]		= {0,2};
-			};
-			class plt_mmap_03_rot
-			{
-				type			= rotate;
-				source			= plt_mmap_r;
-				section			= plt_pr_mpd_back;
-				minValue		= 0;
-				maxValue		= 360;
-				center[]		= { 0.5, 0.75 };
-				angle0			= 0;
-				angle1			= rad 360;
-			};
-			class plt_mmap_04_scale
-			{
-				type = scale;
-				source = plt_mmap_sc;
-				section	= plt_pr_mpd_back;
-				minValue = 0.01;
-				maxValue = 100;
-				center[] = { 0.50, 0.75 };
-				scale0[] = {0.01,0.01};
-				scale1[] = {100,100};
-			};
-			class cpg_mmap_01_tnsx : plt_mmap_01_tnsx
-			{
-				section	= cpg_pr_mpd_back;
-				source = cpg_mmap_tx;
-			};
-			class cpg_mmap_02_tnsy : plt_mmap_02_tnsy
-			{
-				section	= cpg_pr_mpd_back;
-				source = cpg_mmap_ty;
-			};
-			class cpg_mmap_03_rot : plt_mmap_03_rot
-			{
-				section	= cpg_pr_mpd_back;
-				source = cpg_mmap_r;
-			};
-			class cpg_mmap_04_scale : plt_mmap_04_scale
-			{
-				section	= cpg_pr_mpd_back;
-				source = cpg_mmap_sc;
-			};
-		};
+		#include "cfgVehicles\UVAnimations.hpp"
+		#include "cfgVehicles\animationSources.hpp"
 
-		class AnimationSources
-		{
-			#define ANIMS_MMAP(seat) \
-			class seat##_mmap_tx \
-			{ \
-				source = "user"; \
-				animPeriod = 0.001; \
-				initPhase=0; \
-			}; \
-			class seat##_mmap_ty \
-			{ \
-				source = "user"; \
-				animPeriod = 0.001; \
-				initPhase=0; \
-			}; \
-			class seat##_mmap_r \
-			{ \
-				source = "user"; \
-				animPeriod = 0.001; \
-				initPhase=0; \
-			}; \
-			class seat##_mmap_sc \
-			{ \
-				source = "user"; \
-				animPeriod = 0.001; \
-				initPhase=1; \
-			}; 
-		ANIMS_MMAP(plt)
-		ANIMS_MMAP(cpg)
-		////////RADAR///////
-		class fcr_enable
-		{
-			displayName = "Install FCR";
-			author = "Apache mod development team";
-			onPhaseChanged = "_this # 0 enableVehicleSensor [""ActiveRadarSensorComponent"",_this # 1 == 1];";
-			source = "user";
-			initPhase = 1;
-			animPeriod = 0.001;
-		};
-		class msn_equip_american
-		{
-			displayName = "Install American Mission Equipment";
-			author = "Apache mod development team";
-			onPhaseChanged = "";
-			source = "user";
-			initPhase = 1;
-			animPeriod = 0.001;
-		};
-		class msn_equip_british
-		{
-			displayName = "Install British Mission Equipment";
-			author = "Apache mod development team";
-			onPhaseChanged = "";
-			source = "user";
-			initPhase = 0;
-			animPeriod = 0.001;
-		};
-		class mast
-		{
-			source = "user";
-			animPeriod = 0.1;
-			initPhase=1;
-		};
-		class magazine_set_1200
-		{
-			displayName = "Swap IAFS for 1200rnd magazine";
-			author = "Apache mod development team";
-			onPhaseChanged = "_this remoteExec [""fza_fnc_weaponSwapM230Mag"",_this # 0];";
-			source = "user";
-			initPhase = 0;
-			animPeriod = 0.001;
-		};
-		class hstab
-		{
-			source = "user";
-			animPeriod = 0.01;
-			initPhase  = 1;
-		};
-		class pnvs
-		{
-			source = "user";
-			animPeriod = 0.01;
-			initPhase=0;
-		};
-		class pnvs_vert
-		{
-			source = "user";
-			animPeriod = 0.01;
-			initPhase=0;
-		};
-		class tads_stow
-		{
-			source = "user";
-			animPeriod = 1;
-			initPhase  = 1;
-			onPhaseChanged = "[_this # 0, ""fza_ah64_tadsStow"", ""tads_stow""] call fza_fnc_animReset";
-		};
-		class m230_recoil
-		{
-			source = "reload";
-			weapon = "fza_m230";
-		};
-		class mainTurret
-		{
-			source = "user";
-			animPeriod = 0.1;
-			initPhase=0;
-		};
-		class maingun
-		{
-			source = "user";
-			animPeriod = 0.1;
-			initPhase  = 0.298;
-		};
-		class pylon1
-		{
-			source = "user";
-			animPeriod = 0.1;
-			initPhase=0;
-		};
-		class pylon2
-		{
-			source = "user";
-			animPeriod = 0.1;
-			initPhase=0;
-		};
-		class pylon3
-		{
-			source = "user";
-			animPeriod = 0.1;
-			initPhase=0;
-		};
-		class pylon4
-		{
-			source = "user";
-			animPeriod = 0.1;
-			initPhase=0;
-		};
-		class swashplate_up_tns
-		{
-			source = "user";
-			animPeriod = 0.1;
-			initPhase=0;
-		};
-		class swashplate_up_pitch
-		{
-			source = "user";
-			animPeriod = 0.1;
-			initPhase=0;
-		};
-		class swashplate_up_bank
-		{
-			source = "user";
-			animPeriod = 0.1;
-			initPhase=0;
-		};
-		class swashplate_dn_tns
-		{
-			source = "user";
-			animPeriod = 0.1;
-			initPhase=0;
-		};
-		class swashplate_dn_pitch
-		{
-			source = "user";
-			animPeriod = 0.1;
-			initPhase=0;
-		};
-		class swashplate_dn_bank
-		{
-			source = "user";
-			animPeriod = 0.1;
-			initPhase=0;
-		};
-		class mr_act_tns
-		{
-			source = "user";
-			animPeriod = 0.1;
-			initPhase=0;
-		};
-		class swup_arm1
-		{
-			source = "user";
-			animPeriod = 0.1;
-			initPhase=0;
-		};
-		class swup_arm1_t
-		{
-			source = "user";
-			animPeriod = 0.1;
-			initPhase=0;
-		};
-		class swup_arm2
-		{
-			source = "user";
-			animPeriod = 0.1;
-			initPhase=0;
-		};
-		class swup_arm3
-		{
-			source = "user";
-			animPeriod = 0.1;
-			initPhase=0;
-		};
-		class swup_arm3_t
-		{
-			source = "user";
-			animPeriod = 0.1;
-			initPhase=0;
-		};
-		class swup_arm4
-		{
-			source = "user";
-			animPeriod = 0.1;
-			initPhase=0;
-		};
-		class trsw
-		{
-			source = "user";
-			animPeriod = 0.1;
-			initPhase=0;
-		};
-		class tr_blade1_pitch
-		{
-			source = "user";
-			animPeriod = 0.1;
-			initPhase=0;
-		};
-		class tr_blade2_pitch
-		{
-			source = "user";
-			animPeriod = 0.1;
-			initPhase=0;
-		};
-		class tr_blade3_pitch
-		{
-			source = "user";
-			animPeriod = 0.1;
-			initPhase=0;
-		};
-		class tr_blade4_pitch
-		{
-			source = "user";
-			animPeriod = 0.1;
-			initPhase=0;
-		};
-		class blade1_pitch
-		{
-			source = "user";
-			animPeriod = 0.1;
-			initPhase=0;
-		};
-		class blade1_flap
-		{
-			source = "user";
-			animPeriod = 0.1;
-			initPhase=0;
-		};
-		class blade2_pitch
-		{
-			source = "user";
-			animPeriod = 0.1;
-			initPhase=0;
-		};
-		class blade2_flap
-		{
-			source = "user";
-			animPeriod = 0.1;
-			initPhase=0;
-		};
-		class blade3_pitch
-		{
-			source = "user";
-			animPeriod = 0.1;
-			initPhase=0;
-		};
-		class blade3_flap
-		{
-			source = "user";
-			animPeriod = 0.1;
-			initPhase=0;
-		};
-		class blade4_pitch
-		{
-			source = "user";
-			animPeriod = 0.1;
-			initPhase=0;
-		};
-		class blade4_flap
-		{
-			source = "user";
-			animPeriod = 0.1;
-			initPhase=0;
-		};
-		//////weapons///////////
-		class pod_m260_1
-		{
-			source = "user";
-			animPeriod = 0.001;
-			initPhase=0;
-		};
-		class pod_m260_2
-		{
-			source = "user";
-			animPeriod = 0.001;
-			initPhase=0;
-		};
-		class pod_m260_3
-		{
-			source = "user";
-			animPeriod = 0.001;
-			initPhase=0;
-		};
-		class pod_m260_4
-		{
-			source = "user";
-			animPeriod = 0.001;
-			initPhase=0;
-		};
-		class rail_m299_1
-		{
-			source = "user";
-			animPeriod = 0.001;
-			initPhase=0;
-		};
-		class rail_m299_2
-		{
-			source = "user";
-			animPeriod = 0.001;
-			initPhase=0;
-		};
-		class rail_m299_3
-		{
-			source = "user";
-			animPeriod = 0.001;
-			initPhase=0;
-		};
-		class rail_m299_4
-		{
-			source = "user";
-			animPeriod = 0.001;
-			initPhase=0;
-		};
-			////////////////////////////////
-			///////////COCKPIT//////////////
-			////////////////////////////////
-			class plt_eng1_start
-            {
-				source = "user";
-				animPeriod = 0.15;
-				initPhase=0;
-				sound = "eng1_start_sound";
-            };
-			class plt_eng2_start
-            {
-				source = "user";
-				animPeriod = 0.15;
-				initPhase=0;
-				sound = "eng2_start_sound";
-            };
-			class plt_rtrbrake
-            {
-				source = "user";
-				animPeriod = 0.001;
-				initPhase=1;
-				onPhaseChanged = "[_this # 0, ""fza_ah64_rtrbrake"", ""plt_rtrbrake""] call fza_fnc_animReset";
-            };
-			class plt_anticollision
-            {
-				source = "user";
-				animPeriod = 0.001;
-				initPhase=0;
-				onPhaseChanged = "[_this # 0, ""fza_ah64_anticollision"", ""plt_anticollision""] call fza_fnc_animReset";
-            };
-			class plt_firesw
-            {
-				source = "user";
-				animPeriod = 0.001;
-				initPhase=0.5;
-            };
-			class cpg_firesw
-            {
-				source = "user";
-				animPeriod = 0.001;
-				initPhase=0.5;
-            };
-			class plt_nvsmode
-            {
-				source = "user";
-				animPeriod = 0.001;
-				initPhase=0;
-            };
-			class cpg_nvsmode
-            {
-				source = "user";
-				animPeriod = 0.001;
-				initPhase=0;
-            };
-			class plt_batt
-            {
-				source = "user";
-				animPeriod = 0.001;
-				initPhase=0;
-				onPhaseChanged = "[_this # 0, ""fza_ah64_battery"", ""plt_batt""] call fza_fnc_animReset";
-            };
-			class plt_apu
-            {
-				source = "user";
-				animPeriod = 0.001;
-				initPhase=0;
-				onPhaseChanged = "[_this # 0, ""fza_ah64_apu"", ""plt_apu""] call fza_fnc_animReset";
-            };
-			class cpg_ihadss_brt
-            {
-				source = "user";
-				animPeriod = 0.001;
-				initPhase=0;
-            };
-			class plt_eng1_throttle
-            {
-				source = "user";
-				animPeriod = 1;
-				initPhase=0;
-				onPhaseChanged = "[_this # 0, ""fza_ah64_powerLever1"", ""plt_eng1_throttle""] call fza_fnc_animReset";
-            };
-			class plt_eng2_throttle
-            {
-				source = "user";
-				animPeriod = 1;
-				initPhase=0;
-				onPhaseChanged = "[_this # 0, ""fza_ah64_powerLever2"", ""plt_eng2_throttle""] call fza_fnc_animReset";
-            };
-						//////////////MPD OBJECTS//////////////////
-#define SECTION(val) \
-			class val##_x {\
-				source= user;\
-				animPeriod = 0.001;\
-			};\
-			class val##_y {\
-				source=user;\
-				animPeriod = 0.001;\
-			};
-			LIST_OF_BONES(pl)
-			LIST_OF_BONES(pr)
-			LIST_OF_BONES(cl)
-			LIST_OF_BONES(cr)
-		//////////////DOORS//////////////////
-		class pdoor
-		{
-			displayName = "Open Pilot door";
-			author = "Apache mod development team";
-			source = "user";
-			animPeriod = 1;
-			initPhase=0;
-		};
-		class gdoor
-		{
-			displayName = "Open Gunner door";
-			author = "Apache mod development team";
-			source = "user";
-			animPeriod = 1;
-			initPhase=0;
-		};
-		///////////////ADS//////////////////
-		class l_ads_p
-        {
-			source = "user";
-			animPeriod = 0.001;
-			initPhase=0;
-        };
-		class l_ads_b : l_ads_p
-        {
-			source = "user";
-        };
-		class l_ads_y : l_ads_p
-        {
-			source = "user";
-        };
-		class r_ads_p
-        {
-			source = "user";
-			animPeriod = 0.001;
-			initPhase=0;
-        };
-		class r_ads_b : r_ads_p
-        {
-			source = "user";
-        };
-		class r_ads_y : r_ads_p
-        {
-			source = "user";
-        };
-	};
 	class MarkerLights
 	{
 		class PositionWhite
