+++ @AH-64D Apache Longbow/Addons/fza_ah64_controls/config/cfgVehicles/textureSources.hpp
@@ -1,0 +1,95 @@
+class textureSources
+{
+    class b2
+    {
+        displayName = "Default";
+        author = "NodUnit";
+        textures[]= {"\fza_ah64_us\tex\Ex\b2_co.paa","\fza_ah64_us\tex\ex\fcr_co.paa"};
+    };
+    class b2_weather
+    {
+        displayName = "Default (weathered)";
+        author = "Franze";
+        textures[]= {"\fza_ah64_us\tex\Ex\b2_weather_co.paa","\fza_ah64_us\tex\ex\fcr_co.paa"};
+    };
+    class arb229th
+    {
+        displayName = "229th ARB";
+        author = "Apache mod development team";
+        textures[]= {"\fza_ah64_us\tex\Ex\229arb_co.paa","\fza_ah64_us\tex\ex\fcr_co.paa"};
+    };
+    class arb229th_weather
+    {
+        displayName = "229th ARB (weathered)";
+        author = "Apache mod development team";
+        textures[]= {"\fza_ah64_us\tex\Ex\229arb_weather_co.paa","\fza_ah64_us\tex\ex\fcr_co.paa"};
+    };
+    class indian_air_force
+    {
+        displayName = "Indian Air Force";
+        author = "BradMick";
+        textures[]= {"\fza_ah64_us\tex\Ex\b2_india_co.paa","\fza_ah64_us\tex\ex\fcr_co.paa"};
+    };
+    class rnlaf_q_09
+    {
+        displayName = "RNLAF 301 Squadron Q-09";
+        author = "Houseman";
+        textures[]= {"\fza_ah64_us\tex\Ex\b2_rnlaf_q_09_co.paa","\fza_ah64_us\tex\ex\fcr_co.paa"};
+    };
+    class rnlaf_demo
+    {
+        displayName = "RNLAF 301 Squadron Demo Team";
+        author = "Houseman";
+        textures[]= {"\fza_ah64_us\tex\Ex\b2_rnlaf_demo_co.paa","\fza_ah64_us\tex\ex\fcr_co.paa"};
+    };
+    class UK_AAC
+    {
+        displayName = "United Kingdom Army Air Corps";
+        author = "Rosd6(Dryden) & NodUnit";
+        textures[]= {"\fza_ah64_us\tex\Ex\UKAAC.paa","\fza_ah64_us\tex\ex\fcr_co.paa"};
+    };
+    class UK_AAC_weathered
+    {
+        displayName = "United Kingdom Army Air Corps (weathered)";
+        author = "Rosd6(Dryden) & NodUnit";
+        textures[]= {"\fza_ah64_us\tex\Ex\UKAAC_weathered.paa","\fza_ah64_us\tex\ex\fcr_co.paa"};
+    };
+    class UK_AAC_SF
+    {
+        displayName = "United Kingdom Army Air Corps Special Forces";
+        author = "Rosd6(Dryden) & NodUnit";
+        textures[]= {"\fza_ah64_us\tex\Ex\UKAAC_SF.paa","\fza_ah64_us\tex\ex\fcr_co.paa"};
+    };
+    class 211th_clean
+    {
+        displayName = "1/211th ARB ""The Air Pirates"" Utah National Guard";
+        author = "seven10 & Apache mod development team";
+        textures[]= {"\fza_ah64_us\tex\Ex\211th_co.paa","\fza_ah64_us\tex\ex\fcr_co.paa"};
+    };
+    class 211th_weathered
+    {
+        displayName = "1/211th ARB ""The Air Pirates"" Utah National Guard (weathered)";
+        author = "seven10 & Apache mod development team";
+        textures[]= {"\fza_ah64_us\tex\Ex\211th_weather_co.paa","\fza_ah64_us\tex\ex\fcr_co.paa"};
+    };
+    class Israeli_Air_Force
+    {
+        displayName = "Israeli Air Force (IAF Saraf)";
+        author = "NodUnit";
+        textures[]= {"\fza_ah64_us\tex\Ex\Nodunit_Isreal_co.paa","\fza_ah64_us\tex\ex\fcr_co.paa"};
+    };
+    class Shield
+    {
+        displayName = "Shield (Marvel)";
+        author = "NodUnit";
+        textures[]= {"\fza_ah64_us\tex\Ex\Shield.paa","\fza_ah64_us\tex\ex\fcr_co.paa"};
+    };
+    class NodUnit
+    {
+        displayName = "Kane's Wrath";
+        author = "NodUnit";
+        textures[]= {"\fza_ah64_us\tex\Ex\Nodunit.paa","\fza_ah64_us\tex\ex\fcr_co.paa"};
+    };
+};
+textureList[] = {"b2", 1};
+hiddenSelectionsTextures[] = {"\fza_ah64_us\tex\Ex\b2_co.paa","\fza_ah64_us\tex\ex\fcr_co.paa"};
