+++ @AH-64D Apache Longbow/Addons/fza_ah64_systems/cfgVehicles.hpp
@@ -1,0 +1,6 @@
+class CfgVehicles {
+    class Helicopter_Base_F;
+    class fza_ah64base : Helicopter_Base_F {
+        #include "\fza_ah64_systems\cfgVehicles\hitPoints.hpp"
+    };
+};
